PMM-12338 Move logger into global utils folder.
