+++ agent/main.go
@@ -15,15 +15,11 @@
 package main
 
 import (
-	"fmt"
-	"path/filepath"
-	"runtime"
-
-	"github.com/sirupsen/logrus"
 	"gopkg.in/alecthomas/kingpin.v2"
 
 	"github.com/percona/pmm/agent/commands"
 	"github.com/percona/pmm/agent/config"
+	"github.com/percona/pmm/utils/logger"
 	"github.com/percona/pmm/version"
 )
 
@@ -33,22 +29,7 @@ func main() {
 		panic("pmm-agent version is not set during build.")
 	}
 
-	// we don't have configuration options for formatter, so set it once there
-	logrus.SetFormatter(&logrus.TextFormatter{
-		FullTimestamp:   true,
-		TimestampFormat: "2006-01-02T15:04:05.000-07:00",
-
-		CallerPrettyfier: func(f *runtime.Frame) (string, string) {
-			_, function := filepath.Split(f.Function)
-
-			// keep a single directory name as a compromise between brevity and unambiguity
-			dir, file := filepath.Split(f.File)
-			dir = filepath.Base(dir)
-			file = fmt.Sprintf("%s/%s:%d", dir, file, f.Line)
-
-			return function, file
-		},
-	})
+	logger.SetupGlobalLogger()
 
 	// check that command-line flags and environment variables are correct,
 	// parse command, but do try not load config file
