+++ managed/services/management/azure_database.go
@@ -30,7 +30,7 @@ import (
 
 	azurev1beta1 "github.com/percona/pmm/api/managementpb/azure"
 	"github.com/percona/pmm/managed/models"
-	"github.com/percona/pmm/managed/utils/logger"
+	"github.com/percona/pmm/utils/logger"
 )
 
 const (
