+++ managed/services/management/node_mgmt_test.go
@@ -35,9 +35,9 @@ import (
 	"github.com/percona/pmm/api/inventorypb"
 	nodev1beta1 "github.com/percona/pmm/api/managementpb/node"
 	"github.com/percona/pmm/managed/models"
-	"github.com/percona/pmm/managed/utils/logger"
 	"github.com/percona/pmm/managed/utils/testdb"
 	"github.com/percona/pmm/managed/utils/tests"
+	"github.com/percona/pmm/utils/logger"
 )
 
 func TestMgmtNodeService(t *testing.T) {
