+++ managed/utils/interceptors/interceptors.go
@@ -32,7 +32,7 @@ import (
 	"google.golang.org/grpc/status"
 
 	"github.com/percona/pmm/api/agentpb"
-	"github.com/percona/pmm/managed/utils/logger"
+	"github.com/percona/pmm/utils/logger"
 )
 
 func logRequest(l *logrus.Entry, prefix string, f func() error) (err error) {
