+++ qan-api2/main.go
@@ -27,8 +27,6 @@ import (
 	_ "net/http/pprof" //nolint:gosec
 	"os"
 	"os/signal"
-	"path/filepath"
-	"runtime"
 	"strings"
 	"sync"
 	"time"
@@ -56,7 +54,7 @@ import (
 	aservice "github.com/percona/pmm/qan-api2/services/analytics"
 	rservice "github.com/percona/pmm/qan-api2/services/receiver"
 	"github.com/percona/pmm/qan-api2/utils/interceptors"
-	"github.com/percona/pmm/qan-api2/utils/logger"
+	"github.com/percona/pmm/utils/logger"
 	"github.com/percona/pmm/utils/sqlmetrics"
 	"github.com/percona/pmm/version"
 )
@@ -278,22 +276,7 @@ func main() {
 
 	log.Printf("%s.", version.ShortInfo())
 
-	logrus.SetFormatter(&logrus.TextFormatter{
-		FullTimestamp:   true,
-		TimestampFormat: "2006-01-02T15:04:05.000-07:00",
-
-		CallerPrettyfier: func(f *runtime.Frame) (string, string) {
-			_, function := filepath.Split(f.Function)
-
-			// keep a single directory name as a compromise between brevity and unambiguity
-			var dir string
-			dir, file := filepath.Split(f.File)
-			dir = filepath.Base(dir)
-			file = fmt.Sprintf("%s/%s:%d", dir, file, f.Line)
-
-			return function, file
-		},
-	})
+	logger.SetupGlobalLogger()
 
 	if *debugF {
 		logrus.SetLevel(logrus.DebugLevel)
